+++ package.json
@@ -13,7 +13,8 @@
     "lint": "balena-lint -e ts -e js src/ test/ typings/ build-utils/ webpack.config.js",
     "test:build": "tsc --noEmit && tsc --noEmit --project tsconfig.js.json",
     "test:unit": "mocha --config test/unit/.mocharc.js",
-    "test:integration": "mocha --config test/integration/.mocharc.js",
+    "test:integration": "find test/integration -name *.spec.ts | xargs -n 1 -I {} sh -c 'mocha --config test/integration/.mocharc.js {} || exit 255'",
+    "test:integration:single": "find test/integration -name *.spec.ts | xargs mocha --config test/integration/.mocharc.js",
     "test:legacy": "mocha --config test/legacy/.mocharc.js",
     "test:node": "npm run test:unit && npm run test:integration && npm run test:legacy",
     "test:env": "docker-compose -f docker-compose.test.yml -f docker-compose.dev.yml up --build; npm run compose:down",
